move prefcategory to controls
